requirements.txt
appdirs==1.4.4
argh==0.26.2
argon2-cffi==20.1.0
asgiref==3.3.1
asn1crypto==1.4.0
astroid==2.4.2
astropy==4.0.2
async_generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.4
babel==2.8.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bitarray==1.6.1
bkcharts==0.2
bleach==3.2.1
bokeh==2.2.3
boto==2.49.0
bottleneck==1.3.2
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
clyent==1.2.2
colorama==0.4.4
comtypes==1.1.7
contextlib2==0.6.0.post1
cryptography==3.1.1
curl
cycler==0.10.0
cython==0.29.21
cytoolz==0.11.0
dask==2.30.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
distributed==2.30.1
django==3.1.4
docutils==0.16
entrypoints==0.3
et_xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
flake8==3.8.4
flask==1.1.2
freetype-py==1.1
fsspec==0.8.3
future==0.18.2
gevent==20.9.0
glob2==0.7
greenlet==0.4.17
grpcio==1.34.0
h5py==2.10.0
heapdict==1.0.1
html5lib==1.1
icc_rt==2019.0.0
icu
idna==2.10
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==2.0.0
iniconfig==1.1.1
intel-openmp
intervaltree==3.1.0
ipykernel==5.3.4
ipython==7.19.0
ipython_genutils==0.2.0
ipywidgets==7.5.1
isort==5.6.4
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.1
jinja2==2.11.2
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter_client==6.1.7
jupyter_console==6.2.0
jupyter_core==4.6.3
jupyterlab==2.2.6
jupyterlab_pygments==0.1.2
jupyterlab_server==1.2.0
keyring==21.4.0
kiwisolver==1.3.0
lazy-object-proxy==1.4.3
llvmlite==0.34.0
locket==0.2.0
lxml==4.6.1
markupsafe==1.1.1
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl
mkl-service==2.3.0
mkl_fft==1.2.0
mkl_random==1.1.1
mock==4.0.2
more-itertools==8.6.0
mpmath==1.1.0
multipledispatch==0.6.0
mysql-connector-python==8.0.22
navigator-updater==0.2.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
networkx==2.5
nltk==3.5
nose==1.3.7
notebook==6.1.4
numba==0.51.2
numexpr==2.7.1
numpy==1.19.2
numpydoc==1.1.0
olefile==0.46
openpyxl==3.0.5
packaging==20.4
pandas==1.1.3
path==15.0.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.0.1
pip==20.2.4
pkginfo==1.6.1
pluggy==0.13.1
ply==3.11
prometheus_client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.14.0
psutil==5.7.2
py==1.9.0
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.20
pycurl==7.43.0.6
pydocstyle==5.1.1
pyflakes==2.2.0
pygments==2.7.2
pylint==2.6.0
pymongo==3.11.2
pynacl==1.4.0
pyodbc==4.0.30
pyopenssl==19.1.0
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.17.3
pysocks==1.7.1
pytest==6.1.1
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.35.1
setuptools==50.3.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
sip==4.19.13
six==1.15.0
sqlalchemy==1.3.20
sqlparse==0.4.1
statsmodels==0.12.0
